r/aws_ssoadmin_permission_set: Move functions around.
