+++ internal/service/ssoadmin/permission_set.go
@@ -310,3 +310,49 @@ func provisionPermissionSet(ctx context.Context, conn *ssoadmin.SSOAdmin, arn, i
 
 	return nil
 }
+
+const (
+	permissionSetProvisioningStatusUnknown  = "Unknown"
+	permissionSetProvisioningStatusNotFound = "NotFound"
+)
+
+func statusPermissionSetProvisioning(ctx context.Context, conn *ssoadmin.SSOAdmin, instanceArn, requestID string) retry.StateRefreshFunc {
+	return func() (interface{}, string, error) {
+		input := &ssoadmin.DescribePermissionSetProvisioningStatusInput{
+			InstanceArn:                     aws.String(instanceArn),
+			ProvisionPermissionSetRequestId: aws.String(requestID),
+		}
+
+		resp, err := conn.DescribePermissionSetProvisioningStatusWithContext(ctx, input)
+
+		if err != nil {
+			return nil, permissionSetProvisioningStatusUnknown, err
+		}
+
+		if resp == nil || resp.PermissionSetProvisioningStatus == nil {
+			return nil, permissionSetProvisioningStatusNotFound, nil
+		}
+
+		return resp.PermissionSetProvisioningStatus, aws.StringValue(resp.PermissionSetProvisioningStatus.Status), nil
+	}
+}
+
+const (
+	permissionSetProvisioningRetryDelay = 5 * time.Second
+	permissionSetProvisionTimeout       = 10 * time.Minute
+)
+
+func waitPermissionSetProvisioned(ctx context.Context, conn *ssoadmin.SSOAdmin, instanceArn, requestID string) (*ssoadmin.PermissionSetProvisioningStatus, error) {
+	stateConf := retry.StateChangeConf{
+		Delay:   permissionSetProvisioningRetryDelay,
+		Pending: []string{ssoadmin.StatusValuesInProgress},
+		Target:  []string{ssoadmin.StatusValuesSucceeded},
+		Refresh: statusPermissionSetProvisioning(ctx, conn, instanceArn, requestID),
+		Timeout: permissionSetProvisionTimeout,
+	}
+	outputRaw, err := stateConf.WaitForStateContext(ctx)
+	if v, ok := outputRaw.(*ssoadmin.PermissionSetProvisioningStatus); ok {
+		return v, err
+	}
+	return nil, err
+}
